refactor : 참가자의 결과를 딜러가 직접 알려주도록 변경
