+++ src/main/kotlin/blackjack/model/Card.kt
@@ -17,3 +17,4 @@ class Card (
         }
     }
 }
+
