+++ src/main/kotlin/blackjack/model/Players.kt
@@ -14,7 +14,7 @@ class Players(
 
     fun results(dealer: Dealer): Map<String, WinningResult> =
         value.associate { player ->
-            player.name to CalculateResult.getUserResult(dealer, player)
+            player.name to dealer.getUserResult(player)
         }
 
     companion object {
